fix(LogViewer): Adjust parsed data to remove tab character
When we estimate the height in guessRowHeight, the data we are using has a \t character rather than tabs. This leads to a misleading character count, which impacts the calculated number of rows displayed. Some example data was provided by OpenShift that had overlapping rows. This data had tabs in it. When tabs were removed or replaced with strings, the issue could not be recreated. There may be other issues involved here, but this at least fixes this one provided example. We should talk to products that are continuing to experience this and gather sample outputs to see if we can recreate/adjust on an individual basis.